Resume_Match_Prediction_Using_NLP-main/requirements.txt
streamlit==1.38.0
pdfminer.six==20201018
spacy==3.7.2
thinc==8.2.2
numpy==1.23.0
setuptools>=65.5.0
python==3.11
pip==25.0.1
